+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [v0.10.0] - unreleased
 
 ### Added
-* Add to line the ability to find a range of characters or a word under the cursor.
+* Add to line the ability to find a word to complete before the cursor
 * Add :history_size keyword to allow configuration of maximum history size
 * Add the ability to replace a current line in the history buffer
 * Add support for preserving edits to lines stored in history